% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myexpit.R
\name{myexpit}
\alias{myexpit}
\title{inverse logistic link function}
\usage{
myexpit(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
z
}
\description{
a function that returns the inverse of the logistic link function
}
\examples{
myexpit(2)
}
\author{
Aya Mitani
}
